LayoutTests/platform/wpe/imported/w3c/web-platform-tests/pointerevents/pointerevent_touch-action-verification-expected.txt

Pointer Events touch-action attribute support

Test Description: Test will automatically check parsing behaviour of various touch-action combinations.

touch-action: basic verification


PASS default 
FAIL stylesheet-none assert_equals: expected "auto" but got "none"
PASS explicit-auto 
FAIL explicit-pan-x assert_equals: expected "auto" but got "pan-x"
FAIL explicit-pan-y assert_equals: expected "auto" but got "pan-y"
FAIL explicit-pan-x-pan-y assert_equals: expected "auto" but got "pan-x pan-y"
FAIL explicit-pan-y-pan-x assert_equals: expected "auto" but got "pan-x pan-y"
PASS explicit-manipulation 
FAIL explicit-none assert_equals: expected "auto" but got "none"
PASS explicit-invalid-1 
FAIL explicit-invalid-2 assert_equals: expected "auto" but got "none"
PASS explicit-invalid-3 
PASS explicit-invalid-4 
PASS explicit-invalid-5 
PASS explicit-invalid-6 
PASS explicit-invalid-7 
PASS explicit-invalid-8 
PASS explicit-invalid-9 
PASS explicit-invalid-10 
PASS explicit-invalid-11 
PASS explicit-invalid-12 
PASS explicit-invalid-13 
PASS not-inherited 
FAIL inherit assert_equals: expected "auto" but got "none"
PASS initial 
